Move kernel finder
