+++ src/kernels/finder/kernelFinder.web.ts
@@ -3,13 +3,13 @@
 
 import { injectable, inject, named } from 'inversify';
 import { Memento } from 'vscode';
-import { GLOBAL_MEMENTO, IMemento } from '../platform/common/types';
-import { IJupyterRemoteCachedKernelValidator, IJupyterServerUriStorage, IServerConnectionType } from './jupyter/types';
+import { GLOBAL_MEMENTO, IMemento } from '../../platform/common/types';
+import { IJupyterRemoteCachedKernelValidator, IJupyterServerUriStorage, IServerConnectionType } from '../jupyter/types';
 import { BaseKernelFinder } from './kernelFinder.base';
-import { PreferredRemoteKernelIdProvider } from './jupyter/preferredRemoteKernelIdProvider';
-import { IRemoteKernelFinder } from './raw/types';
-import { INotebookProvider, KernelConnectionMetadata } from './types';
-import { IApplicationEnvironment } from '../platform/common/application/types';
+import { PreferredRemoteKernelIdProvider } from '../jupyter/preferredRemoteKernelIdProvider';
+import { IRemoteKernelFinder } from '../raw/types';
+import { INotebookProvider, KernelConnectionMetadata } from '../types';
+import { IApplicationEnvironment } from '../../platform/common/application/types';
 
 /**
  * Web version of a KernelFinder. Web has different ways to validate than node.
